refactor(lotto): Relocate the classes in lotto.domain package to make same with production codes
