+++ src/test/kotlin/lotto/domain/LottoGameTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.LottoGame
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.DisplayName
