+++ src/test/kotlin/lotto/domain/LottoPrizeStaticsTest.kt
@@ -1,10 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
-import lotto.domain.WinningLotto
-import lotto.domain.LottoNumber
-import lotto.domain.LottoPrizeStatics
-import lotto.domain.Prize
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
